Add presenter for abstract morphomath operation process.

+++ src/layers/medComposer/process/morphomath_operation/medMorphomathOperationProcessNode.tpp
@@ -18,12 +18,12 @@
 #include <dtkLog>
 
 #include <medAbstractImageData.h>
-#include <medDoubleParameter.h>
+#include <medIntParameter.h>
 
 class medMorphomathOperationProcessNodePrivate
 {
 public:
-    dtkComposerTransmitterReceiver<double> radius;
+    dtkComposerTransmitterReceiver<int> radius;
     dtkComposerTransmitterReceiver<medAbstractImageData*> input;
 
     dtkComposerTransmitterEmitter<medAbstractImageData*> output;
